TCMGreensFunctions

Documentation for TCMGreensFunctions.

TCMGreensFunctions.BoundaryPropagator

TCMGreensFunctions.TracerInverseGaussian

TCMGreensFunctions.boundary_propagator_timeseries

TCMGreensFunctions.boundary_propagator_timeseries

TCMGreensFunctions.boundary_propagator_timeseries

TCMGreensFunctions.BoundaryPropagator — Type

BoundaryPropagator

Modeling boundary propagation with propagator functions and atmospheric source terms.

Fields

Gp_arr: Matrix of boundary propagator functions

f_atm: Vector of atmospheric source functions

t_vec: Vector of time points

C0: Initial tracer concentration vector (optional)

t0: Reference time (optional)

TCMGreensFunctions.TracerInverseGaussian — Type

TracerInverseGaussian(Γ,Δ)

using LinearAlgebra: NumberArray

The tracer inverse Gaussian distribution with mean Γ and width Δ has probability density function

\[G(𝐱, \tau) = \sqrt{\frac{\Gamma^3 }{4 \pi \Delta^2 \tau^3 }} \exp \left( - \frac{\Gamma (\tau - \Gamma)^2}{4 \Delta ^2 \tau}\right) \]

TracerInverseGaussian()              # Tracer Inverse Gaussian distribution with unit mean and unit width, i.e. TracerInverseGaussian(1, 1)
TracerInverseGaussian(Γ, Δ)          # Tracer Inverse Gaussian distribution with mean Γ and width Δ

params(d)           # Get the parameters, i.e. (Γ, Δ)
mean(d)             # Get the mean parameter, i.e. Γ
shape(d)            # Get the shape parameter, i.e. Δ

External links

Compare to Inverse Gaussian distribution on Wikipedia

TCMGreensFunctions.boundary_propagator_timeseries — Method

boundary_propagator_timeseries(Gp_arr, f_atm, t_vec)

Calculate convolutions for matrix of propagators with vector of sources.

Returns 3D array with dimensions [propagatorrow, sourcecolumn, time].

TCMGreensFunctions.boundary_propagator_timeseries — Method

boundary_propagator_timeseries(bp::BoundaryPropagator)

Calculate convolution using data stored in a BoundaryPropagator object.

TCMGreensFunctions.boundary_propagator_timeseries — Method

boundary_propagator_timeseries(Gp, f_atm, t_vec)

Calculate convolution of a boundary propagator with surface source.